requirements.txt
fastai
gensim
gradio
ipywidgets
jupyter
nltk
notebook
pandas
plotly
pytest
scikit-learn
seaborn
setuptools
simpletransformers
spacy
transformers
wheel
scipy
python-dotenv
torch 
torchvision 
torchaudio
tensorflow